+++ drivers/sensor/qdec_nxp_s32/qdec_nxp_s32.c
@@ -10,7 +10,7 @@
 #include <zephyr/drivers/pinctrl.h>
 #include <zephyr/drivers/sensor.h>
 #include <zephyr/logging/log.h>
-#include <zephyr/sys/arch_interface.h>
+#include <zephyr/arch/arch_interface.h>
 
 #ifndef M_PI
 #define M_PI 3.14159265358979323846
