+++ include/zephyr/timing/timing.h
@@ -7,7 +7,7 @@
 #ifndef ZEPHYR_INCLUDE_TIMING_TIMING_H_
 #define ZEPHYR_INCLUDE_TIMING_TIMING_H_
 
-#include <zephyr/sys/arch_interface.h>
+#include <zephyr/arch/arch_interface.h>
 #include <zephyr/timing/types.h>
 
 #ifdef __cplusplus
