+++ samples/kernel/condition_variables/simple/src/main.c
@@ -6,7 +6,6 @@
 
 #include <zephyr/kernel.h>
 #include <zephyr/arch/cpu.h>
-#include <zephyr/sys/arch_interface.h>
 
 #define NUM_THREADS 20
 #define STACK_SIZE (1024)
