+++ subsys/shell/modules/device_service.c
@@ -12,7 +12,7 @@
 #include <zephyr/device.h>
 #include <zephyr/pm/device.h>
 #include <zephyr/pm/device_runtime.h>
-#include <zephyr/sys/arch_interface.h>
+#include <zephyr/arch/arch_interface.h>
 
 static const char *get_device_name(const struct device *dev,
 				   char *buf,
