+++ tests/arch/common/timing/src/main.c
@@ -5,7 +5,7 @@
  */
 #include <zephyr/ztest.h>
 #include <zephyr/kernel.h>
-#include <zephyr/sys/arch_interface.h>
+#include <zephyr/arch/arch_interface.h>
 
 #if defined(CONFIG_SMP) && CONFIG_MP_MAX_NUM_CPUS > 1
 #define SMP_TEST
